Extract css injection logic
