+++ src/common/content/expander/index.js
@@ -4,6 +4,8 @@ const extend = require('xtend')
 const gyazoIdFromUrl = require('./lib/gyazoIdFromUrl')
 const adjacentStyle = require('./lib/adjacentStyle')
 
+require('./injectCss')()
+
 function fetchImage (url, callback) {
   chrome.runtime.sendMessage(chrome.runtime.id, {
     action: 'gyazoGetImageBlob',
@@ -23,41 +25,6 @@ function fetchImage (url, callback) {
     xhr.send()
   })
 }
-
-const node = document.createElement('style')
-node.innerHTML = `
-.gz-circle-loader {
-  display: inline-block;
-  text-indent: -9999em;
-  margin: 0;
-  border-top: 8px solid rgba(0, 0, 0, .2);
-  border-right: 8px solid rgba(0, 0, 0, .2);
-  border-bottom: 8px solid rgba(0, 0, 0, .2);
-  border-left: 8px solid #000;
-  animation: spin 1.1s infinite linear;
-
-  border-radius: 50%;
-  width: 32px;
-  height: 32px;
-
-  &:after {
-    border-radius: 50%;
-    width: 32px;
-    height: 32px;
-  }
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-`
-
-document.body.appendChild(node)
 
 function createLoader (position = {}) {
   const loader = document.createElement('div')
